% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapUEB_function.R
\name{HeatMapUEB}
\alias{HeatMapUEB}
\title{Plot Heatmap}
\usage{
HeatMapUEB(data, title, cexTitle, scale)
}
\arguments{
\item{data}{A dataset where only the first column is character (where to put your GeneSymbols)}

\item{title}{A short name to put in the title after "Heatmap of "}

\item{cexTitle}{Size of the title}

\item{scale}{Does the data need to be scaled? Values= "column"/"row"/"none"}
}
\description{
This function plots a Heapmap of a datamatrix where the first column are the genesymbols
}
\examples{
HeatMapUEB(data=hmdata,"Normalized Data",0.8,scale="column")
}

